% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/visualize.R
\name{populateMatrix}
\alias{populateMatrix}
\title{Helper function for main visualize function}
\usage{
populateMatrix(query, subject)
}
\arguments{
\item{query}{Query}

\item{subject}{Subject}
}
\description{
Helper function for main visualize function
}
\examples{
populateMatrix()
}

