% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/visualize.R
\name{visualize}
\alias{visualize}
\title{Visualization Function}
\usage{
visualize(gff, clust_count, file = "plot.png")
}
\arguments{
\item{gff}{The file path to the gff path storing all
the transcripts of interest.}

\item{clust_count}{The number of clusters to sort into.}

\item{file}{The output png file path. Defaults to "plot.png"}
}
\description{
This function automatically clusters dna transcripts
and visualizes them by cluster
}
\examples{
visualize()
}

